Update to uint32_t.

+++ Erigo/Inc/cmd_msg_struct.h
@@ -5,9 +5,9 @@
 
 typedef struct WAV_CMD_DATA{
 	uint8_t cmd_id;
-    uint16_t test_amp_ma;
-	uint16_t nm_amp_ma;
-	uint16_t freq_sel;
+    uint32_t test_amp_ma;
+	uint32_t nm_amp_ma;
+	uint32_t freq_sel;
 }WAV_CMD_DATA;
 
 #endif
